Fixing text support. Fixes #9.
After a lot of trial and error, the method that works best to render text
is to specify a default font size of 1 unit, and then use the T2 transform of
the diagrams text primitive to add an SVG transform attribute.

One problem with this approach is that stroke parameters get transformed in
wonky ways leading to bad outputs. The workaround for that is to set the stroke
explicitly to "none" in the <text> node.

The other thing to note is that the default fill color is "transparent", so we
need to explicitly specify a fill color for text nodes if we are not specifying
fill color elsewhere. Otherwise there will be no visible output. Cairo does not
have this behavior, so we need to tweak all Cairo examples to add an explicit
fill color to text primitives. Filed #15 to track this.