Fix for clipping with freeze support.
Adding freeze support broke clipping.  The clip path needs to have the
same transformations applied as a normal path.  But as an attribute it
gets rendered through a different code path.  So we apply the inverse of
the freeze transform to get it back in line with a normal path.