Fixed issue with stacked clipping.
Before this several stacked clips were just printed as several paths in the clipPath element.
But according to the SVG standard this means they are OR'ed together, although we want them to be intersected.
I changed this by building them up around the clipped element hierarchical, which leads to intersection according to the SVG spec.
I tested this in the chart drawing tests and it seems to work.